.history/requirements_20251224090748.txt
pyautogui
keyboard
img2pdf
pygetwindow
Pillow
